Forward declare `composite_blur_filter_data` to sort out visibility warning.
Seems there was a bit of circular include happening here.

Clang was producing this all over:

```
src/blur/gaussian.h:16:46: declaration of 'struct composite_blur_filter_data' will not be visible outside of this function [-Werror,-Wvisibility]
```

We can forward declare `composite_blur_filter_data` and move the `box.h` and `gaussian.h` to the source file to avoid the circular include.

It looks like `composite_blur_filter_data` can be moved to the source file but i've left that alone for now.