Add shims for libobs private utils from `dstr.h` and `platform.h`
libobs only ships a subset of public headers (see `libobs/CMakeLists.txt` for the list)
in install bundles. `dstr` and `platform.h` are both missing, so I've just copied in the headers and replaced `EXPORT` with extern that way the linker will use the libobs
implementation. In the case of `dstr` there are a lot of inline functions so that has an unfortunate amount of overlap with the libobs version.

Not sure this is the best way forward, but it at least gets things building.
It's probably unsustainable to depend on the private obs headers across versions
given this will be prone to ABI / API breakage.

Some other options:
- These are useful enough that upstream may want to expose them to plugin authors, but that's a longer term route.
- Fork the utils needed by just copying in from libobs. Now you maintain those implementations.
- Write own impl of these utils

If it were me, I'd rewrite these in a few lines of c++ :)